bug -listreplay and -listlevels didn't work

+++ src/Renderer.cpp
@@ -1166,8 +1166,8 @@ namespace vapp {
 	      beta = acos(-p[i].y / v_ray) - M_PI / 2.0;
 	    }
 	    
-	    p[i].x = cos(pSprite->DrawAngle() + beta) * v_ray;
-	    p[i].y = sin(pSprite->DrawAngle() + beta) * v_ray;
+	    p[i].x = (int)(cos(pSprite->DrawAngle() + beta) * v_ray);
+	    p[i].y = (int)(sin(pSprite->DrawAngle() + beta) * v_ray);
 	  }
 	  //pSprite->setDrawAngle(pSprite->DrawAngle() + 0.01);
 	}
